Define __STDC_VERSION__ as long constant

+++ tests/fixtures/builtin_version.c
@@ -1,3 +1,3 @@
-int foo() {
+long foo() {
     return __STDC_VERSION__;
 }
